requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
aiomysql==0.3.2
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.2.2
pyjwt==2.12.1
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.27
redis==7.4.0